id should be Long

+++ springboot_backend/src/main/java/com/example/springbootbackend/controller/UserController.java
@@ -19,7 +19,7 @@ public class UserController {
     }
 
     @DeleteMapping("/{id}")
-    public void delete(@PathVariable int id){
+    public void delete(@PathVariable Long id){
         userService.delete(id);
     }
